+++ src/app/page.js
@@ -1,13 +1,14 @@
-import Navbar from '@/components/Navbar'
-import Image from 'next/image'
+import Navbar from '@/components/nav/Navbar'
+import Image from '@/components/Image'
+import king from '../../public/images/logos/King-Fuyuki-pfp.jpg'
 
 export default function Home() {
   return (
     <div className="flex flex-row items-center justify-between p-0 m-0 box-border">
       
       <p>hello</p>
-      <p className="bg-red-600">hi</p>
+      <Image imgURL={king.src} />
       <p>dev branch</p>
     </div>
-  )
+  );
 }
